Fix bug in "previous sentence logic"
We used an array-map for storing the tagged sentences, but did not keep
in mind that the insertion order gets lost on subsequent modifications.
This had the effect of returning sentences older than the previously
tagged on upon clicking the button.
We now use a regular map for storing the tagged sentences, but keep an
additional vector of sentence-ids to keep the tagging order around.